backend/requirements.txt
# Backend REST API - Sistema de Seguridad IoT
Flask==3.0.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0

# Base de datos
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# MQTT
paho-mqtt==1.6.1

# Validación
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0

# Utilidades
python-dotenv==1.0.0
gunicorn==21.2.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
